Resolve circular dependencies in external grader library
