+++ apps/prairielearn/src/lib/assessment.sql
@@ -8,18 +8,6 @@ WHERE
   ai.id = $assessment_instance_id
   AND a.id = $assessment_id;
 
--- BLOCK select_assessment_for_grading_job
-SELECT
-  ai.id AS assessment_instance_id
-FROM
-  grading_jobs AS gj
-  JOIN submissions AS s ON (s.id = gj.submission_id)
-  JOIN variants AS v ON (v.id = s.variant_id)
-  JOIN instance_questions AS iq ON (iq.id = v.instance_question_id)
-  JOIN assessment_instances AS ai ON (ai.id = iq.assessment_instance_id)
-WHERE
-  gj.id = $grading_job_id;
-
 -- BLOCK select_assessment_info
 SELECT
   assessment_label (a, aset),
